Reduced synchronous steps in the import process and introduced new hooks.
Added two new hooks (IPhotoSource::init and IPhotoSource::final) to
initialize and finalize the photo sources before the ImporterDialog
is shown and after it is hidden. These are meant to be asynchronous
operations. eg., one can connect to IPhotoSource::init_end to be
notified when initialization has ended.

It is now possible to avoid keeping the browser widget as a part of
the IPhotoSource implementations. The widget can be dynamically
created in IPhotoSource::get_browser and Gtk::manage can be used to
delegate its lifetime to that of its parent. A new hook has been
introduced as IPhotoSource::read_selection to read the URIs selected
in the browser after a response has been received from the
ImporterDialog and before it is destroyed.

Importer::on_action_photo_import has been split into the following
three parts to avoid blocking the main loop for too long:
+ To invoke IPhotoSource::init.
+ To show the ImporterDialog on IPhotoSource::init_end.
+ To respond to ImporterDialog::signal_response.