Moved the Zoomer's functionality to the main Gtk::Toolbar
Replace the Zoomer with a ScaleAction which uses a ScaleToolItem as
the proxy. Therefore we can move the slider to the main toolbar for a
more consistent code and UI, and save some space. We no longer need
the "tree" icons from F-Spot as the ScaleToolItem does not have any.

A new placeholder called PlaceholderMiscToolBar was added to allow
right-justified Gtk::ToolItems to be added.