Introduced Tracker and the thumbnail management specification
This is a pretty big change affecting the entire meta-data storage
and thumbnailing code. Instead of using our own SQLite database for
keeping our meta-data we now use Tracker [1] so that the meta-data
(eg., tags) is no longer locked in Solang but can be accessed across
the desktop. SparQL is used to talk to Tracker over DBus. For
thumbnailing we use the the DBus Specification for Thumbnail
Management [2]. Any thumbnailer service which provides the standard
org.freedesktop.thumbnails.xxx API can be used. Tumbler [3] is one
such implementation.

One of the major fallouts of this change has been that most of the
blocking interfaces have now been replaced by their asynchronous
cousins. Those that are left either need to be replaced or should not
be used. This is mainly to avoid blocking DBus calls [4].

Since it is possible to query Tracker to return all photos known to
it, there is no need to manually import photos. The only scenario
where this will not work is if Tracker's indexer has been switched
off. In that case one would have to explicitly inform Tracker about
them. Even then, something like Conduit [5] might be a good option.
Therefore the Importer and all implementations of IPhotoSource have
been disabled.

The storage subsystem has been removed. Due to the above changes it
was no longer needed to have one and henceforth we will rely on Gvfs
for our needs. All implementations of IStorage have been removed.

ExifData fields are now populated from the information provided by
Tracker according to the Nepomuk Multimedia (NMM) ontology.

Since we are using the Thumbnail Management Specification, thumbnails
are no longer separately created and stored in
$XDG_DATA_HOME/solang/thumbnails. Instead we follow the Thumbnail
Managing Standard [6].

Some files have been moved from src/attribute and src/database to
src/common to satisfy the linker.

Dependencies:

Exiv2 and libgdamm have been dropped. Tracker-client and dbus-glib
are the new compile time dependencies. The presence of the
org.freedesktop.Tracker1 and org.freedesktop.thumbnails.xxx DBus
services are the new runtime requirements.

Regressions:

The editor is broken.

The DateManager has been disabled until the queries are ported to
work with Tracker.

Tags only have a name and can not be removed from photos.

[1] http://tracker-project.org/
[2] http://live.gnome.org/ThumbnailerSpec
[3] http://git.xfce.org/apps/tumbler/
[4] http://smcv.pseudorandom.co.uk/2008/11/nonblocking/
[5] http://live.gnome.org/Conduit
[6] http://jens.triq.net/thumbnail-spec/