Improved thumbnail handling
Do not keep the thumbnail (and other textual data) in
Application::listStore_. Just keep the serial number and PhotoPtr, and
use GtkCellLayoutDataFuncs to set the Gtk::CellRenderers in the
ThumbnailView. New thumbnails are created using
CellRendererThumbnail::create_thumbnail, which is invoked from
CellRendererThumbnail::render_vfunc. This is done to reduce the number
of thumbnails to be created to only those which will actually get
rendered. As a result of this, BrowserRenderer::generate_thumbnails
has been replaced with BrowserRenderer::set_thumbnail_size.

The Thumbnailer is now a singleton and a new signal
Thumbnailer::ready has been added to indicate when it has completed
processing some photos. Application::on_thumbnailer_ready triggers
the "row-changed" signal on those rows of Application::listStore_
which have any of the processed photos.