Disabled the EditorRenderer
We will try to use the EnlargedRenderer for editing purposes.

Since the BrowserRenderer will be the only one presented when the
application starts, it will not be placed in a Gtk::Notebook by
GDL. This breaks the current mechanism for finding the
BrowserRenderer::pageNum_ and EnlargedRenderer::pageNum_. Now we do
it cleanly by listening to the "parent-set" signal of the
GdlDockItems.