Moved the UI controls away from the PaginationBar to other locations
The PaginationBar no longer provides any UI elements. It only deals
with providing the abstraction of having multiple pages within the
browser and managing movement from one page to another. The previous
and next buttons are now part of the main Gtk::Toolbar, while the
string describing the current page is now shown in the main
Gtk::Statusbar. This makes the code and UI more consistent and saves
some space.