Replaced the Gtk::Buttons of the SearchBasket with a context menu
It would be good to support free text searches by allowing the user to
type in some arbitrary string and trying to retrieve photos with some
matching criteria. In order to make space for a Gtk::Entry in the
SearchBasket, the Gtk::Buttons have been removed. Instead a context
menu has been added which allows selecting all the filters, or
removing the selected filters.