Fixed the flip and rotate operations
Flipping or rotating a photo changes the extent of the GeglBuffer
associated with the photo, and depending on the kind of filter that was
used the extent is stretched by a few pixels. To fix this, the photo
should be translated and cropped after being flipper or rotated. The
gegl:translate and gegl:crop are almost as quick as gegl:nop, and
should not be a significant overhead.

IOperation::get_node now takes a BufferPtr to figure out the
parameters for the gegl:translate and gegl:crop operations.